server/requirements.txt
asgiref
django
django-allauth
djangorestframework
djangorestframework_simplejwt
django-cors-headers
django-extensions
djoser
psycopg
psycopg-binary
redis
requests
python-dotenv
#PyJWT
#social-auth-app-django
#social-auth-core
sqlparse
tzdata
pillow
faker
stripe
qrcode
reportlab
celery
channels
redis